+++ hub/src/main/java/com/faster/hub/app/hub/application/usecase/HubService.java
@@ -1,12 +1,14 @@
 package com.faster.hub.app.hub.application.usecase;
 
-import com.faster.hub.app.hub.application.dto.SaveHubApplicationRequestDto;
-import com.faster.hub.app.hub.application.dto.SaveHubApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.GetHubApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.DeleteHubApplicationRequestDto;
-import com.faster.hub.app.hub.application.dto.UpdateHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.dto.request.GetPathApplicationRequestDto;
+import com.faster.hub.app.hub.application.dto.request.SaveHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.dto.response.GetPathsApplicationResponseDto;
+import com.faster.hub.app.hub.application.dto.response.SaveHubApplicationResponseDto;
+import com.faster.hub.app.hub.application.dto.response.GetHubApplicationResponseDto;
+import com.faster.hub.app.hub.application.dto.request.DeleteHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.dto.request.UpdateHubApplicationRequestDto;
 import java.util.UUID;
-import com.faster.hub.app.hub.application.dto.UpdateHubApplicationResponseDto;
+import com.faster.hub.app.hub.application.dto.response.UpdateHubApplicationResponseDto;
 
 public interface HubService {
