+++ hub/src/main/java/com/faster/hub/app/hub/presentation/HubController.java
@@ -5,13 +5,13 @@ import com.common.resolver.annotation.CurrentUserInfo;
 import com.common.resolver.dto.CurrentUserInfoDto;
 import com.common.resolver.dto.UserRole;
 import com.common.response.ApiResponse;
-import com.faster.hub.app.hub.application.dto.DeleteHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.dto.request.DeleteHubApplicationRequestDto;
 import com.faster.hub.app.hub.application.usecase.HubService;
-import com.faster.hub.app.hub.presentation.dto.GetHubResponseDto;
-import com.faster.hub.app.hub.presentation.dto.SaveHubRequestDto;
-import com.faster.hub.app.hub.presentation.dto.SaveHubResponseDto;
-import com.faster.hub.app.hub.presentation.dto.UpdateHubRequestDto;
-import com.faster.hub.app.hub.presentation.dto.UpdateHubResponseDto;
+import com.faster.hub.app.hub.presentation.dto.response.GetHubResponseDto;
+import com.faster.hub.app.hub.presentation.dto.request.SaveHubRequestDto;
+import com.faster.hub.app.hub.presentation.dto.response.SaveHubResponseDto;
+import com.faster.hub.app.hub.presentation.dto.request.UpdateHubRequestDto;
+import com.faster.hub.app.hub.presentation.dto.response.UpdateHubResponseDto;
 import jakarta.validation.Valid;
 import java.time.LocalDateTime;
 import java.util.UUID;
